ugc_service/requirements.txt
aiokafka~=0.12.0
fastapi~=0.111.0
fastapi-jwt-auth~=0.5.0
pydantic-settings~=2.5.2
